Introduce default fetcher
